Add basic authentication to RestTemplate and update mock tests
In order to secure the API calls, the RestClient in BeerClientMockTest.java is updated to include an "Authorization" header. The RestTemplateBuilderConfig is also modified to accommodate basic authentication using a provided username and password.

The changes ensure that all requests sent from our client to the server are authorized. The mock tests are updated accordingly to expect this Authorization header.