Add OAuth2 client support for REST template
Added OAuth2 client dependency, configuration and interceptor to support OAuth2 authorization. Modified the tests and pom.xml to include the new dependency. The configuration 'Basic' for RestTemplate has been changed to OAuth2. The reasons for this change would be mainly security - using OAuth 2.0, the protocol used industry-wide for API security.